pyproject.toml
[project]
name = "fastapi-admin"
version = "0.1.0"
description = "FastAPI Admin"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.136.1",
    "greenlet>=3.4.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
]
